[LateralMenu] Add Base and Members mixins for lateral menu
